requirements.txt
pandas
numpy
scikit-learn
python-dotenv
pymongo
certifi
pydantic
python-box
pyyaml
xgboost
mlflow
uv